fix(ui): satisfy sidebar CI gates
